@@ -5,7 +5,7 @@ version: 1.0.0
 homepage: https://github.com/full-of-fire/flutter_reachability
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.20.0"
 
 dependencies:
